+++ egui_web/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added ⭐
+
+* Auto-save of app state to local storage
+
 ### Changed ⭐
 
 * Set a maximum canvas size to alleviate performance issues on some machines